Refactored the whole `BasePlayer` hierarchy
